requirements.txt
# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0

# HTTP Requests for API calls
requests==2.31.0
urllib3==2.1.0

# Data handling
pandas==2.1.4
numpy==1.26.2

# Caching
redis==5.0.1
python-redis-cache==3.0.0

# Date/Time handling
python-dateutil==2.8.2
pytz==2023.3

# Data validation
pydantic==2.5.3

# Job scheduling
APScheduler==3.10.4

# Logging
loguru==0.7.2

# Environment variables
python-dotenv==1.0.0

# JSON handling
orjson==3.9.10

# API documentation
flask-swagger-ui==4.11.1
